is_current_alert - move to utils and test it
Test that it recognises both a current and a previous alert based
on expiry date.
