+++ requirements_for_test.txt
@@ -2,4 +2,5 @@
 
 isort==5.8.0
 flake8==3.9.1
+freezegun==1.1.0
 pytest==6.2.3
